requirements.txt
# Streamlit
streamlit==1.33.0

# PDF loaders
PyMuPDF==1.23.3
pypdf==3.17.4

# Core LangChain ecosystem

langchain-text-splitters==0.3.8
langchain-huggingface==0.1.2

langchain==0.3.25
langchain-core==0.3.64
langchain-community==0.3.22

langsmith==0.3.45
pydantic>=2.7.4,<3.0


# Compatible with langchain-core 0.3.64
pydantic-settings==2.4.0

# Vector DB + Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.6.1
huggingface-hub==0.23.0

# LLMs
transformers==4.40.0
torch==2.3.1 # Keep this as it's a good modern version
sentencepiece==0.2.0
tokenizers==0.19.1

# Utilities
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
requests==2.31.0
tqdm==4.66.1
typing-extensions==4.11.0
orjson==3.9.15
pillow==10.2.0
langdetect==1.0.9